feat(dsp): add direct DCT-II implementation and fix test-infrastructure bug
- Added `speech::dsp::dctII` for direct DCT-II computation of arbitrary length N.
- Fixed a test-infrastructure bug where `makeSine()` helpers computed phase in `float`, leading to precision loss at large sample indices and causing spurious MFCC test failures.
- Updated checklist.md to reflect completed DSP layer features and mark MFCC as implemented.
- Added documentation for the new DCT-II implementation and its use in MFCC.

+++ tests/dsp/test_fft.cpp
@@ -10,7 +10,7 @@ namespace {
 std::vector<float> makeSine(int n, float cyclesOverN) {
     std::vector<float> out(n);
     for (int i = 0; i < n; ++i) {
-        out[i] = std::sin(2.0f * static_cast<float>(M_PI) * cyclesOverN * i / n);
+        out[i] = std::sin(2.0 * M_PI * cyclesOverN * i / n);
     }
     return out;
 }
